Reorganized server files
